# perf(sdk/elixir): optimize function call time